requirements.txt
streamlit
lxml
requests
PyMuPDF
google-cloud-vision
google-cloud-aiplatform
futures
PyPDF2
langchain-community
langchain
langchain-core
langchain-google-genai
langchain-google-vertexai
python-docx